+++ packages/spacecat-shared-http-utils/test/index.test.js
@@ -12,9 +12,18 @@
 /* eslint-env mocha */
 import { expect } from 'chai';
 import {
-  ok, badRequest, notFound, internalServerError,
-  noContent, found, created, createResponse, unauthorized, forbidden,
   accepted,
+  badRequest,
+  createResponse,
+  created,
+  forbidden,
+  found,
+  internalServerError,
+  methodNotAllowed,
+  noContent,
+  notFound,
+  ok,
+  unauthorized,
 } from '../src/index.js';
 
 async function testMethod(response, expectedCode, expectedBody) {
@@ -163,6 +172,23 @@ describe('HTTP Response Functions', () => {
     expect(response.headers.get('custom-header')).to.equal('value');
     const responseBody = await response.json();
     expect(responseBody).to.deep.equal({ message: 'Resource not found' });
+  });
+
+  it('methodNotAllowed should return a 405 Method Not Allowed response with default message and headers', async () => {
+    const response = await methodNotAllowed();
+    expect(response.status).to.equal(405);
+    expect(response.headers.get('x-error')).to.equal('method not allowed');
+    const responseBody = await response.json();
+    expect(responseBody).to.deep.equal({ message: 'method not allowed' });
+  });
+
+  it('methodNotAllowed should return a 405 Method Not Allowed response with custom message and headers', async () => {
+    const response = await methodNotAllowed('This method is not allowed', { 'custom-header': 'value' });
+    expect(response.status).to.equal(405);
+    expect(response.headers.get('x-error')).to.equal('This method is not allowed');
+    expect(response.headers.get('custom-header')).to.equal('value');
+    const responseBody = await response.json();
+    expect(responseBody).to.deep.equal({ message: 'This method is not allowed' });
   });
 
   it('internalServerError should return a 500 Internal Server Error response with default message and headers', async () => {