+++ packages/spacecat-shared-utils/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@adobe/spacecat-shared-utils-v1.38.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.38.4...@adobe/spacecat-shared-utils-v1.38.5) (2025-05-31)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#779](https://github.com/adobe/spacecat-shared/issues/779)) ([07f8cce](https://github.com/adobe/spacecat-shared/commit/07f8cce73e33bfb9c61fe14f2ef28012b872437d))
+
+# [@adobe/spacecat-shared-utils-v1.38.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.38.3...@adobe/spacecat-shared-utils-v1.38.4) (2025-05-24)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#762](https://github.com/adobe/spacecat-shared/issues/762)) ([2eb5a22](https://github.com/adobe/spacecat-shared/commit/2eb5a22935a7293ef293157e0b0fe6506d5b4c94))
+
 # [@adobe/spacecat-shared-utils-v1.38.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.38.2...@adobe/spacecat-shared-utils-v1.38.3) (2025-05-19)